Configure workspaces for vitest running
