+++ src/services/api/package.json
@@ -31,6 +31,7 @@
     "build": "tsc",
     "buildLayers": "tsc -p layers",
     "lint": "eslint '**/*.{ts,js}'",
-    "compileRepack": "tsc -p repack"
+    "compileRepack": "tsc -p repack",
+    "test": "vitest"
   }
 }
