+++ src/services/email/package.json
@@ -10,7 +10,8 @@
         "@aws-sdk/client-ses": "^3.499.0"
     },
     "scripts": {
-        "lint": "eslint '**/*.{ts,js}'"
+        "lint": "eslint '**/*.{ts,js}'",
+        "test": "vitest"
     },
     "devDependencies": {
         "@typescript-eslint/eslint-plugin": "^6.20.0",
