+++ src/services/ui/src/components/Cards/index.ts
@@ -1,2 +1,3 @@
 export * from "./CardWithTopBorder";
 export * from "./SectionCard";
+export * from "./OptionCard";
