+++ src/services/ui/testing/setup.ts
@@ -1,6 +1,6 @@
-import { expect, afterEach } from 'vitest'
-import { cleanup } from '@testing-library/react'
-import matchers from '@testing-library/jest-dom/matchers'
+import { expect, afterEach } from "vitest";
+import { cleanup } from "@testing-library/react";
+import * as matchers from "@testing-library/jest-dom/matchers";
 
 // extends Vitest's expect method with methods from react-testing-library
 expect.extend(matchers);
